// finance_client/src/app/user/data/types/customer-types/category.ts
export interface Category {
  totalItems?: number;
  totalPages?: number;
  currentPage?: number;
  items?: [
    {
      id?: 70306;
      segmentname?: string;
      familycode?: string;
      familyname?: string;
      classcode?: string;
      classname? :string;
      commoditycode?: string;
      commodityname?: string;
      isservice?: string;
    }
  ];
}